Treat file length
