+++ feature/Messages/Sources/Messages/Views/MessagesTabView/ConversationRow/Badge.swift
@@ -1,0 +1,25 @@
+//
+//  Badge.swift
+//
+//
+//  Created by Nityananda Zbil on 10.11.23.
+//
+
+import SwiftUI
+
+struct Badge: View {
+    let unreadCount: Int
+
+    var body: some View {
+        if unreadCount > 0 {
+            Text("\(unreadCount)")
+                .foregroundColor(.white)
+                .font(.headline)
+                .padding(.m)
+                .background(.red)
+                .clipShape(Circle())
+        } else {
+            EmptyView()
+        }
+    }
+}
